backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
python-dotenv==1.1.1
pyautogui==0.9.54
pillow==11.3.0
pytesseract==0.3.13
pyperclip==1.9.0
chromadb==1.0.15
pydantic==2.11.7
pandas==2.2.3
numpy==1.26.4
docker==7.1.0
GitPython==3.1.43
psutil==6.1.0
pyppeteer==1.0.2